+++ sql/queries/postgresql.sql
@@ -155,6 +155,10 @@ INSERT INTO credbound.password_credentials (user_id, hash, updated_at) VALUES ($
 -- name: GetPassword :one
 SELECT user_id, hash, updated_at FROM credbound.password_credentials WHERE user_id = $1;
 
+-- name: LockPassword :one
+SELECT user_id, hash, updated_at FROM credbound.password_credentials WHERE user_id = $1
+FOR UPDATE;
+
 -- name: ReplacePassword :execrows
 UPDATE credbound.password_credentials SET hash = $2, updated_at = $3 WHERE user_id = $1;